feat(ui): migrate raw object view to uploads page and clean up unused stored page
